backend/django-backend/pyproject.toml
[project]
name = "jobmatchskr"
version = "0.1.0"
description = ""
authors = [
    {name = "calebba",email = "bawacaleb14@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-crispy-forms (>=2.3,<3.0)",
    "crispy-bootstrap5 (>=2024.10,<2025.0)",
    "django-allauth (>=65.5.0,<66.0.0)",
    "python-environ (>=0.4.54,<0.5.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
